refactor: extract pane osc helpers
